+++ src/Tunnel.Cli/Program.cs
@@ -26,6 +26,7 @@ rootCommand.AddCommand(new StatusCommand().Build());
 rootCommand.AddCommand(new RemoveCommand().Build());
 rootCommand.AddCommand(new ReconnectCommand().Build());
 rootCommand.AddCommand(new CleanCommand().Build());
+rootCommand.AddCommand(new ReloadCommand().Build());
 rootCommand.AddCommand(new UpdateCommand().Build());
 rootCommand.AddCommand(new VersionCommand().Build());